requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-text-splitters
python-dotenv
chromadb
pydantic
faiss-cpu 
sentence-transformers 
langchain-community
langchain-google-genai
